Intermediate commit for Vulkan on Android: video decoding works, video rendering does not.

+++ TeleportServer/src/TeleportServer/GeometryEncoder.h
@@ -18,7 +18,7 @@ namespace teleport
 		~GeometryEncoder() = default;
 
 		// Inherited via GeometryEncoderBackendInterface
-		avs::Result encode(uint32_t timestamp, avs::GeometrySourceBackendInterface* target
+		avs::Result encode(uint64_t timestamp, avs::GeometrySourceBackendInterface* target
 						   , avs::GeometryRequesterBackendInterface* geometryRequester) override;
 		avs::Result mapOutputBuffer(void*& bufferPtr, size_t& bufferSizeInBytes) override;
 		avs::Result unmapOutputBuffer() override;
